Issue/#56

See merge request nerko/nerko-api-test-platform/sms-satp!43

+++ src/main/java/com/sms/satp/service/impl/CaseTemplateServiceImpl.java
@@ -1,11 +1,17 @@
 package com.sms.satp.service.impl;
 
+import static com.sms.satp.common.enums.OperationModule.CASE_TEMPLATE;
+import static com.sms.satp.common.enums.OperationType.ADD;
+import static com.sms.satp.common.enums.OperationType.DELETE;
+import static com.sms.satp.common.enums.OperationType.EDIT;
 import static com.sms.satp.common.exception.ErrorCode.ADD_CASE_TEMPLATE_ERROR;
 import static com.sms.satp.common.exception.ErrorCode.DELETE_CASE_TEMPLATE_ERROR;
 import static com.sms.satp.common.exception.ErrorCode.EDIT_CASE_TEMPLATE_ERROR;
 import static com.sms.satp.common.exception.ErrorCode.GET_CASE_TEMPLATE_PAGE_ERROR;
 import static com.sms.satp.common.exception.ErrorCode.SEARCH_CASE_TEMPLATE_ERROR;
 
+import com.sms.satp.common.aspect.annotation.Enhance;
+import com.sms.satp.common.aspect.annotation.LogRecord;
 import com.sms.satp.common.exception.ApiTestPlatformException;
 import com.sms.satp.dto.PageDto;
 import com.sms.satp.dto.request.AddCaseTemplateRequest;
@@ -25,6 +31,7 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -52,6 +59,7 @@ public class CaseTemplateServiceImpl implements CaseTemplateService {
     }
 
     @Override
+    @LogRecord(operationType = ADD, operationModule = CASE_TEMPLATE, template = "{{#addCaseTemplateRequest.name}}")
     public Boolean add(AddCaseTemplateRequest addCaseTemplateRequest) {
         log.info("CaseTemplateService-add()-params: [CaseTemplate]={}", addCaseTemplateRequest.toString());
         try {
@@ -66,6 +74,8 @@ public class CaseTemplateServiceImpl implements CaseTemplateService {
     }
 
     @Override
+    @LogRecord(operationType = DELETE, operationModule = CASE_TEMPLATE, template = "{{#result?.![#this.name]}}",
+        enhance = @Enhance(enable = true, primaryKey = "ids"))
     public Boolean deleteByIds(List<String> ids) {
         log.info("CaseTemplateService-deleteById()-params: [id]={}", ids);
         try {
@@ -82,6 +92,7 @@ public class CaseTemplateServiceImpl implements CaseTemplateService {
     }
 
     @Override
+    @LogRecord(operationType = EDIT, operationModule = CASE_TEMPLATE, template = "{{#updateCaseTemplateRequest.name}}")
     public Boolean edit(UpdateCaseTemplateRequest updateCaseTemplateRequest) {
         log.info("CaseTemplateService-edit()-params: [CaseTemplate]={}", updateCaseTemplateRequest.toString());
         try {
@@ -136,16 +147,20 @@ public class CaseTemplateServiceImpl implements CaseTemplateService {
 
     private void deleteCaseTemplateApi(List<CaseTemplateApi> caseTemplateApiList) {
         List<String> ids = caseTemplateApiList.stream().map(CaseTemplateApi::getId).collect(Collectors.toList());
-        caseTemplateApiService.deleteByIds(ids);
+        if (CollectionUtils.isNotEmpty(ids)) {
+            caseTemplateApiService.deleteByIds(ids);
+        }
     }
 
     private void editCaseTemplateApiStatus(CaseTemplate caseTemplate, Boolean oldRemove) {
         List<CaseTemplateApi> caseTemplateApiList = caseTemplateApiService
             .getApiByCaseTemplateId(caseTemplate.getId(), oldRemove);
-        for (CaseTemplateApi caseTemplateApi : caseTemplateApiList) {
-            caseTemplateApi.setRemoved(caseTemplate.getRemoved());
+        if (CollectionUtils.isNotEmpty(caseTemplateApiList)) {
+            for (CaseTemplateApi caseTemplateApi : caseTemplateApiList) {
+                caseTemplateApi.setRemoved(caseTemplate.getRemoved());
+            }
+            caseTemplateApiService.editAll(caseTemplateApiList);
         }
-        caseTemplateApiService.editAll(caseTemplateApiList);
     }
 
 }
