+++ src/main/java/com/sms/satp/service/impl/SceneCaseServiceImpl.java
@@ -1,5 +1,9 @@
 package com.sms.satp.service.impl;
 
+import static com.sms.satp.common.enums.OperationModule.SCENE_CASE;
+import static com.sms.satp.common.enums.OperationType.ADD;
+import static com.sms.satp.common.enums.OperationType.DELETE;
+import static com.sms.satp.common.enums.OperationType.EDIT;
 import static com.sms.satp.common.exception.ErrorCode.ADD_SCENE_CASE_ERROR;
 import static com.sms.satp.common.exception.ErrorCode.DELETE_SCENE_CASE_ERROR;
 import static com.sms.satp.common.exception.ErrorCode.EDIT_SCENE_CASE_CONN_ERROR;
@@ -9,6 +13,8 @@ import static com.sms.satp.common.exception.ErrorCode.GET_SCENE_CASE_PAGE_ERROR;
 import static com.sms.satp.common.exception.ErrorCode.SEARCH_SCENE_CASE_ERROR;
 
 import com.google.common.collect.Lists;
+import com.sms.satp.common.aspect.annotation.Enhance;
+import com.sms.satp.common.aspect.annotation.LogRecord;
 import com.sms.satp.common.exception.ApiTestPlatformException;
 import com.sms.satp.dto.PageDto;
 import com.sms.satp.dto.request.AddSceneCaseRequest;
@@ -38,6 +44,7 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -74,6 +81,7 @@ public class SceneCaseServiceImpl implements SceneCaseService {
     }
 
     @Override
+    @LogRecord(operationType = ADD, operationModule = SCENE_CASE, template = "{{#addSceneCaseRequest.name}}")
     public Boolean add(AddSceneCaseRequest addSceneCaseRequest) {
         log.info("SceneCaseService-add()-params: [SceneCase]={}", addSceneCaseRequest.toString());
         try {
@@ -88,8 +96,10 @@ public class SceneCaseServiceImpl implements SceneCaseService {
     }
 
     @Override
+    @LogRecord(operationType = DELETE, operationModule = SCENE_CASE, template = "{{#result?.![#this.name]}}",
+        enhance = @Enhance(enable = true, primaryKey = "ids"))
     public Boolean deleteByIds(List<String> ids) {
-        log.info("SceneCaseService-deleteById()-params: [ids]={}", ids);
+        log.info("SceneCaseService-deleteById()-params: [ids]={}", ids.toString());
         try {
             for (String id : ids) {
                 sceneCaseRepository.deleteById(id);
@@ -104,6 +114,7 @@ public class SceneCaseServiceImpl implements SceneCaseService {
     }
 
     @Override
+    @LogRecord(operationType = EDIT, operationModule = SCENE_CASE, template = "{{#updateSceneCaseRequest.name}}")
     public Boolean edit(UpdateSceneCaseRequest updateSceneCaseRequest) {
         log.info("SceneCaseService-edit()-params: [SceneCase]={}", updateSceneCaseRequest.toString());
         try {
@@ -181,15 +192,15 @@ public class SceneCaseServiceImpl implements SceneCaseService {
     public Boolean editConn(UpdateSceneTemplateRequest updateSceneTemplateRequest) {
         log.info("SceneCaseService-editConn()-params: [SceneTemplateDto]={}", updateSceneTemplateRequest.toString());
         try {
-            if (!updateSceneTemplateRequest.getSceneCaseApiDtoList().isEmpty()) {
+            if (!updateSceneTemplateRequest.getUpdateSceneCaseApiRequests().isEmpty()) {
                 sceneCaseApiService.batchEdit(
                     BatchUpdateSceneCaseApiRequest.builder()
-                        .sceneCaseApiRequestList(updateSceneTemplateRequest.getSceneCaseApiDtoList()).build());
+                        .sceneCaseApiRequestList(updateSceneTemplateRequest.getUpdateSceneCaseApiRequests()).build());
             }
-            if (!updateSceneTemplateRequest.getCaseTemplateConnDtoList().isEmpty()) {
+            if (!updateSceneTemplateRequest.getUpdateCaseTemplateConnRequests().isEmpty()) {
                 List<CaseTemplateConn> caseTemplateConnList =
                     caseTemplateConnMapper
-                        .toCaseTemplateConnList(updateSceneTemplateRequest.getCaseTemplateConnDtoList());
+                        .toCaseTemplateConnList(updateSceneTemplateRequest.getUpdateCaseTemplateConnRequests());
                 caseTemplateConnService.editList(caseTemplateConnList);
             }
             return Boolean.TRUE;
@@ -210,8 +221,10 @@ public class SceneCaseServiceImpl implements SceneCaseService {
 
     private void deleteSceneCaseApi(String id) {
         List<SceneCaseApi> sceneCaseApiList = sceneCaseApiService.listBySceneCaseId(id);
-        List<String> ids = sceneCaseApiList.stream().map(SceneCaseApi::getId).collect(Collectors.toList());
-        sceneCaseApiService.deleteByIds(ids);
+        if (CollectionUtils.isNotEmpty(sceneCaseApiList)) {
+            List<String> ids = sceneCaseApiList.stream().map(SceneCaseApi::getId).collect(Collectors.toList());
+            sceneCaseApiService.deleteByIds(ids);
+        }
     }
 
     private void deleteCaseTemplate(String id) {
@@ -224,10 +237,12 @@ public class SceneCaseServiceImpl implements SceneCaseService {
     private void editSceneCaseApiStatus(SceneCase sceneCase, Boolean oldRemove) {
         List<SceneCaseApi> sceneCaseApiList = sceneCaseApiService
             .getApiBySceneCaseId(sceneCase.getId(), oldRemove);
-        for (SceneCaseApi sceneCaseApi : sceneCaseApiList) {
-            sceneCaseApi.setRemoved(sceneCase.getRemoved());
+        if (CollectionUtils.isNotEmpty(sceneCaseApiList)) {
+            for (SceneCaseApi sceneCaseApi : sceneCaseApiList) {
+                sceneCaseApi.setRemoved(sceneCase.getRemoved());
+            }
+            sceneCaseApiService.editAll(sceneCaseApiList);
         }
-        sceneCaseApiService.editAll(sceneCaseApiList);
     }
 
 }
